style : feedList 하나의 컴포넌트로 변경
