+++ src/pages/MainPage.tsx
@@ -1,32 +1,11 @@
 import { useNavigate } from 'react-router-dom';
 
 import CircleButton from '@/components/CircleButton/CircleButton';
-import Feed, { FeedType } from '@/components/feed/Feed';
-import Lnb from '@/components/lnb/Lnb';
+import FeedContainer from '@/components/feedcontainer/FeedContainer';
 import Nav from '@/components/navigation/Nav';
 
 const MainPage = () => {
     const navigate = useNavigate();
-    const feedArr: FeedType[] = [
-        {
-            index: 0,
-            title: '제목을 쓰는 공간입니다.',
-            heart: 7,
-            comment: 4,
-            content:
-                '이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은',
-            date: '2024.01.01. 12:00',
-        },
-        {
-            index: 1,
-            title: '제목을 쓰는 공간입니다.',
-            heart: 7,
-            comment: 4,
-            content:
-                '이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은 내용을 쓰는 공간입니다. 이곳은',
-            date: '2024.01.01. 12:00',
-        },
-    ];
 
     return (
         <>
@@ -53,10 +32,8 @@ const MainPage = () => {
                 ]}
                 onUrl={function iu() {}}
             />
-            <Lnb />
-            {feedArr.map((data) => (
-                <Feed key={data.index} data={data} />
-            ))}
+            {/* 피드 리스트 모아보기  */}
+            <FeedContainer />
             <CircleButton
                 iconName={'i-edit-fill'}
                 iconColor='#fff'
